np_helper: guard size_t multiplications and zero-size reductions
* transpose.c: NPdtranspose_021 and NPztranspose_021 computed
  "size_t nm = shape[1] * shape[2]" in int and only widened on
  assignment. Cast first operand to size_t so matrices larger than
  46340 x 46340 don't silently overflow.

* pack_tril.c: NP{d,z}{unpack,pack}_tril_2d had the same int-times-int
  pattern for "size_t nn = n * n" and "size_t n2 = n*(n+1)/2".

* condense.c: NP_Bmax, NP_imax, NP_fmax read a[0] unconditionally and
  could read past the legitimate slice when called with di == 0 or
  dj == 0 (which NPbcondense / NPicondense / NPfcondense will do for
  loc_x[i] == loc_x[i+1] groups). Add the same di == 0 || dj == 0
  guard that NP_max / NP_min / NP_absmax / NP_absmin / NP_norm
  already have.

+++ pyscf/lib/np_helper/transpose.c
@@ -56,7 +56,7 @@ void NPdtranspose_021(int *shape, double *a, double *at)
         shared(shape, a, at)
 {
         int ic;
-        size_t nm = shape[1] * shape[2];
+        size_t nm = (size_t)shape[1] * shape[2];
 #pragma omp for schedule (static)
         for (ic = 0; ic < shape[0]; ic++) {
                 NPdtranspose(shape[1], shape[2], a+ic*nm, at+ic*nm);
@@ -70,7 +70,7 @@ void NPztranspose_021(int *shape, double complex *a, double complex *at)
         shared(shape, a, at)
 {
         int ic;
-        size_t nm = shape[1] * shape[2];
+        size_t nm = (size_t)shape[1] * shape[2];
 #pragma omp for schedule (static)
         for (ic = 0; ic < shape[0]; ic++) {
                 NPztranspose(shape[1], shape[2], a+ic*nm, at+ic*nm);
